requirements.txt
asgiref==3.6.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.0.1
click==8.1.3
cloudinary==1.32.0
colorama==0.4.6
confection==0.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
cymem==2.0.7
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.2
django-cors-headers==3.14.0
django-debug-toolbar==4.1.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
filelock==3.12.0
identify==2.5.23
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.2.0
langcodes==3.3.0
MarkupSafe==2.1.2
murmurhash==1.0.9
nodeenv==1.7.0
numpy==1.24.3
oauthlib==3.2.2
packaging==23.0
pandas==2.0.1
pathy==0.10.1
pdfminer==20191125
pdfminer.six==20221105
phonenumbers==8.13.11
Pillow==9.5.0
platformdirs==3.5.0
pre-commit==3.2.2
preshed==3.0.8
psycopg2-binary==2.9.7
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
smart-open==6.3.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.3
srsly==2.4.6
thinc==8.1.9
tqdm==4.65.0
typer==0.7.0
typing-extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
uuid==1.30
virtualenv==20.22.0
wasabi==1.1.1